Replaced aliased ffi type and added platform check
Replaced the aliased type from mxChar, which was a `std::ffi::short`
with the rust native type `u16`.

Added a compile time assertion to check if the compilation target is 64-bit.
This is necessary to ensure correctness of the bindings. Maybe this
requirement could be relaxed in the future. Since the last version of
Matlab available in 32-bit was 2015b the restriction to 64-bit shouldn't
be to bad for most cases.