Reworked exported API
Grouped bindings for `mxClassID` and `mxComplexity`, which were previously
freestanding constants, into their own module.

Removed some undocumented symbols from the exported items. A search through
all examples shipped with Matlab did not reveal a single use for the removed items.

Adjusted the xtask `generate-bindings` to remove version suffixes while
generating the bindings. This removes the error prone string manipulation
that solved this task previously while reexporting. Adjusted the passed
options for the changes mentioned above.

Adjusted the xtask `reexport-versionless` to not care about the version suffixes
any longer. Also added a output file command line argument for convenience.
The reexports are now harder to regenerate because the current algorithm
does not handle the nested module well. Since the exported API surface should
not change very much if at all this is an acceptable tradeoff from the current
point of view.

Fixed up the mex-examples, which were slightly broken because of the grouping
of the `mxClassID` and `mxComplexity` constants into their modules. Ran tests
to validate.